arbint* -> arbint
Before this commit, the basic data type was the `arbint`, defined as

    typedef struct {
        uint32_t* value,
        enum sign sign,
        size_t length
    } arbint;

However, it soon became clear that it was more efficient to pass around
pointers to such a struct instead of a copy of the struct. This lead to
a lot of address-of and dereference operators lingering around in the
public API*, which was ugly and confusing. To remedy this, I hereby
change the basic data type to be a pointer to a struct like above. That
way the user can only handle the pointers, while the dereferencing and
memory management is only done in the backend (see include/datatypes.h
of this commit).

* yet to be defined